+++ program.h
@@ -18,6 +18,5 @@ int program_find_bootable_partition(struct list_head *ops, bool *multiple_found)
 int program_is_sec_partition_flashed(struct list_head *ops);
 int program_cmd_add(struct list_head *ops, const char *address, const char *filename);
 int erase_cmd_add(struct list_head *ops, const char *address);
-int program_resolve_gpt_deferrals(struct qdl_device *qdl, struct list_head *ops);
 
 #endif
